services/sync-messaging-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
websockets==12.0
redis==5.0.1
celery==5.3.4
schedule==1.2.1
cryptography==41.0.8
aiofiles==23.2.1
pydantic-settings==2.1.0
email-validator==2.1.0
pillow==10.1.0
python-magic==0.4.27
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0
